style(frontend): moved imports and files around
also a better way to do the lookups in the types
