+++ frontend/src/lib/components/modal/pokemonDetailModal.svelte
