+++ frontend/src/lib/components/modal/typeCardSmall.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-	import type { TypesType } from '$lib/types';
 	import { PokeSearch } from '$lib/state/pokeSearch.svelte';
 	import { PokeSelect } from '$lib/state/pokeSelect.svelte';
+	import type { TypesType } from '$lib/types';
 
 	let { type } = $props<{ type: TypesType }>();
